simrun: avoid useless computation with state.regs.ip
If ip is not symbolic, there is no need to ask the solver to get
solutions, and then to create a constraint, and store again the ip
into the new state.

Also, use target instead of state.regs.ip whenever possible, to
avoid calling a load() in memory.py since we already have the value.