Add Concrete FS mode to Posix plugin
When this mode is enabled the filesystem the simulation is being run on
will be respected. Calls to open and fopen will return error codes if
the file they requested does not exist on the host's filesystem.
Additionally if the file does exist the file backing will be concrete
(SimConcreteFile) opposed to being symbolically backed with a
SimSymbolicFile