Capture the case when the temporary variable for next_expr is missing.
Under certain cases we don't want to fully execute the entire basic
block (for the sake of better performance),for example, you can create a
slice of a program, but only execute the first half of the last basic
block (and suppose the tmp used in `next_expr` of that basic block is
not included in the slice). In old days, since the tmp is not included
in the slice, a KeyError will be directly thrown out to the user, and
the final state is gone. Now the final state is stored to
`unsat_successors` even if the tmp that `default_exit` relies on doesn't
exist.