Remove concrete loading for CGC binaries
angr now has unicorn support, so this should now be made unnecessary.
Unicorn emulation may be a bit slower, but it should take care of
a lot of state tracking (tracking reads/writes for example) that
we failed to do with the QEMU implementation.

Additionally, in this commit we add a symbolic page to represent the
secret CGC flag data.