Remove CapstoneInsn and CopyClass in lifter since they are not needed.
According to file capstone/__init__.py in the Python binding of
capstone, a one-to-one copy of _cs_insn is made by copy_ctypes().
Although the original `_cs_insn` array is freed right after the call to
cs_disasm(), content in the copy is managed by Python, and is not freed
immediately. Hence it should be safe to directly use capstone classes.