Make DDG.simplified_data_graph a multi graph to support killer edges.
Consider the following instruction:
    inc eax, 1
This instruction will produce a killer edge and a consuming edge for the
eax register. If the simplified_data_graph is not a multi graph, one of
those two edges are missing.

It's not a problem for the non-simplified data dependence graph since
the value of eax is first read into a temporary variable, and then used
and written to the eax register. Read and write occur at different code
locations.