Find/Avoid conflict suggested fix
If either find or avoid is a user-supplied boolean function I do not know how to resolve conflicts.
As a solution I added a boolean parameter 'avoid_priority'.
Its default value is False to mimick the current behavior.

If both are addresses, we want to choose the first to occur.
I had to modify __init__.py: when the first address of a run is in the find (or avoid) set the corresponding function now returns a set containing that address instead of a boolean.
It was necessary in order to be able to distinguish between this case (first address is valid) and user-supplied functions.
I ran the tests in angr/tests, everything run fine.