CFGFast: do not force-update the function address of a CFGNode.
In certain cases a CFGNode might be assigned a function address that is
different than the function address used in _generate_cfgnode() the
second time it is traced. We do not force update the function address of
this CFGNode in this case, rather, we update the current function
address of the analysis. If the CFGNode's function address is incorrect,
hopefully it will be corrected during post-processing.