CFGFast: Fix a rare case leading to incorrect function-returning decisions. (#604)
* CFGFast: Fix a rare case leading to incorrect function-returning decisions.

In CFGFast, we assume any function call should return to the same function
as its caller block as long as the callee returns. This holds in general,
but it does not hold in the following scenario during CFG recovery:

- We identify a function starting at address A. This function returns.
- We identify another function starting at address B. Unfortunately, many
  blocks of function B and function A overlap. This might be the result of
  multi-entry functions (in which case, both are correct), or function A
  starts at the alignment bytes ahead of function B (in which case,
  function B is the correct one, and function A should only contain the
  alignment bytes and a jumpout to the beginning of function B).
- At this time, all blocks other than the first block in function B belong
  to function A. Therefore, any transition going out of the first block of
  function B are outside transitions (they all transition to function A).

I noticed this issue a long time ago (probably before CGC), and added a
fix. However, I was only handling the cases when the first block of
function B jumps to its successors. The case where the first block calls
another function and then returns was not handled. This commit fixes this
issue.

Function 804a420 in the CGC binary CROMU_00071 can be used to test this
commit. It is included in Patcherex as a test case:
test_cfg.py:test_fullcfg_properties.

* Skip all retout sites whose callee returning status are undetermined.