CFGFast: Make popping pending jobs much faster. (#883)
* CFGFast: Make popping pending jobs much faster.

When job queue is empty, we used to unnecessarily check the returning
status of each function in self.kb.functions if there is a pending job
associated to it, which is really slow when there are many pending jobs
with many functions (it becomes really noticeable when > 5000). This
commit implements a PendingJobs class that records all functions that have
been marked as returning, so we don't have to query the returning state of
every function before popping a pending job.