More polishing on the plugin refactor (#897)
* More polishing on the plugin refactor!

* Get rid of the circular dependencies between SimState and SimStatePlugin

- Make state plugins register themselves directly with the SimState.register_default(), thus eliminating the need in _hub_type;
- Remove plugin imports from the state.py;
- SimState should obtain the 'memory' and the 'registers' plugins by making request to the active preset, instead of directly instantiating the SimSymbolicMemory, SimFastMemory, etc. classes. This brings more flexibility in deciding what exact memory implementation should be used when o.ABSTRACT_MEMORY, o.FAST_MEMORY, etc. options are engaged.
- Leave it for plugins to decide on how to store a state, when the SimStatePlugin.set_state() is called. For example, SimAbstractMemory should store a strong reference to a state, while other plugins should not.

* Make EngineHub more flexible

First of all, this commit removes the _choose_engine() method from the EngineHubPreset class. Presets should definitely _not_ contain any kind of processing logic. Instead, they should provide a list of default plugins, and, possibly, a number of other adjustments, that the hub should incorporate in order to work with the provided plugins properly.

Also, the current implementation of engine chooser limits the processing
order to ['failure', 'syscall', 'hook', other specified in the preset],
thus not allowing one to include any user-specified engines, registered
via the register_plugin(), in that order without altering the preset. This is not to mention that it is impossible to execute any engine before the 'failure' engine or others. I would say, the latter is crucial to me.

As an alternative to that, I suggest using the `EngineHub.order` property
to define the order, in which engines to be enumerated. The `EngineHubPreset.order` property makes a complement to this by specifying
in which order its plugins should be executed.

Second, this commit brings back the `default_engine` and the `procedure_engine` properties to the EngineHub, as well, as the `EngineHub.sucessors()` arguments under the same names. This allow one to specify any engine to go as default/procedure engine, without being bound to any particular engine name. This also allows one to use EngineHub.successors() processing logic without being bound by the current processing order that is set on the hub. 

Last, and not the least, this commit "un-inherits" Plugin base class from SimEngine class. It's natural, that the SimEngine seems like a plugin
from the EngineHub perspective but, essentialy, they're not plugins.
Of course, they could be plugged into the EngineHub, but other than that,
they're stand-alone objects that could be used on their own. There're a lot of places in angr, where SimEngine is used without being accessed from what was the AngrFactory back then. So, for me, it seems like a bit of misunderstaing of what a plugin (and an engine) actually is.

* Replace Analysis.register_default() with AnalysesHub.register_default()

Same as for the SimState plugin

* Remove Plugin class as it is no more needed

The less entities, the better!

* Move 'insn_text' handling from EngineHub to SimEngineVEX

EngineHub should not care about the arguments that will be passed to the SimEngine.process() method

* Linting

* EngineHub: add `predefined_order` initialization argument

* PluginPreset: implement copy() method

* EnginePreset: implement copy() method

* Rename NoPlugin to AngrNoPluginError

* Add 'procedure' engine to engine basic preset

* EngineHub: check if default/procedure engines exist before trying to get them

* PluginHub: use list_default_plugins() instead of list_plugins()

* Linting

* Make hook_symbol work correctly when the symbol is weak and auto_load_libs=False

* Fix spiller test case for Appveyor. (#911)

* Fix RST errors in docstrings