Lifter: Fix the crash caused by lifting a block with unfilled holes in memory under fastpath mode. Closes #1062. (#1069)
When a state under fastpath mode is given, it will throw a
SimSolverModeError upon calling eval() on the block bytes if the block is
not continuous, i.e., there are holes of missing memory bytes that must be
filled with unconstrained symbolic variables. This commit addresses this
issue by catching SimErrors and falling back to the slow path after the
exception occurs.