CFGFast: Fix a bug that may cause missing jumpout sites.
In some cases, CFGFast might fail to set the jumpout sites of a function
if:
- The function does not have any return instructions, and
- The function has a tail jump that jumps to another function that
  returns.

This is because in the second phase of CFG recovery, we are directly
copying the returning setting of all functions that has been previously
found (during the first phase). If a function has been incorrectly marked
as not returning during the first phase, we will then inherit the false
positive in the second phase.

The obvious solution is to only inherit true positives (functions that
must return) from the first analysis phase.