First attempt to fix bug 1476.
The inspect plugin previously followed the following cases for handling event attributes
in the check method:
1. If both current and needed attributes are none, return true.
2. If current xor needed attributes are none, return false.
3. If current attribute is a solution to needed attribute return true.
4. Else, return false.

The problem is that not all of the attributes are solver objects;
some are strings, or possibly other values.

This fix adds a step before step 3:
- If current == needed, return true.

Also, it modifies step 3 to only accept instances of claripy.base.Base
and SimActionObjects objects.