Added unit test for inspect mem_read_fallback property.
This is a simple example demonstrating the capability with an "always on"
handler.  Selectively activating this handler would allow for
things like isolating the values of arguments that are passed by reference,
which can't be captured by generaly by inspecting the call site
because the load length isn't presented uniformly.