Refactors get_unconstrained_bytes() out
Also reproduces the uninit warning(s) in SimSymbolicMemory in
FullySymbolisMemory

Adds SAT check for uninit read in _load(); the paper has this but the
reference implementation did not (due to perf?). It's no less correct,
but this way we reduce log spam.

Uses DUMMY_SYMBOLIC_READ_VALUE instead of 0 in _load()

Fixes a latent bug in _store()