ReachingDefinitions: refactor `ReachingDefinitions`
Constructor took `Block` *AND* `Function` as parameter, but could never
analyse both.
Use an enum-like kind-of logic: a single parameter, that can be either a
`Block` or a `Function` and have separated behavior based on which it
is.

Motivation: cleaner code, and easier to add "structure" to analyse (such
as slices...).