CFGFast: Resolve indirect jumps before following undetermined pending jobs.
It was assumed in CFGFast that it will determine the returning status of
all (currently discovered) functions before popping a pending job.
However, it is not the case if determining the returning status of the
function will require resolving an indirect jump first, which is the case
in issue #1844.

This commit solve this specific case by making sure indirect jump
resolving happens prior to the unconfirmed pending jobs are popped out
from the PendingJobs queue.

This commit closes issue #1844.