claripy: ccall: properly extract overflow register
Prior we were xoring (flags & o_mask), which meant we were xoring the
wrong bit. This code was based on earlier adc code, which only dealt
with cf, so (flags & c_mask) already had the bit in the lowest location
by coincidence. However, since of is not in the lowest location, we have
to properly shift. So, this changes the logic for both cf and of to
shift and then mask by 1.

It also more properly passes through untouched registers and fixes up
the tests.

Fixes: #1937
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>