Fix ordering of node addrs for _libc_start_main
As symbols are identified and processed first, _libc_start_main
can be encountered before the entry node in certian binaries.
this causes the procedure to be run on the wrong entry block addr,
leading to an incorrect parsing of the init/main/fini arguments
and subsequent missing blocks from the generated cfg.

Reordering the starting_points ensures the correct entry addr is
used as the src_node when _lib_start_main procedure is called