Handle tailcalls when making return edges with detect_tail_calls set
_make_return_edges only checks immediate predecessors for return addrs
when determining successors for return edges. This skips functions
that are tailcalls, as the caller it returns to is by definition
not the immediate predecessor, leading to an inaccurate cfg.

Since tailcall detection is already done under cfg_base, added a set
to track tailcall addrs found when detect_tail_calls is set, so
implementations like cfg_fast can reuse this information.

fixed linting issues