SimOS: Warn about the use of stdin when bytes are provided.
When stdin is provided, angr defaults to treating it as the entire
content of stdin and does not allow reading beyond the end of the
content. This commit adds a warning to users so that they may manually
set has_end=False when they intend to provide the prefix of the stdin
instead of the entire stdin.