Interleave concretized addresses to speed up solving.
Building cascading Or expressions on a list of sorted ints may significantly
slow down constraint solving since it is much easier for the solver to hit
corner cases, such as having to attempt all sub-expressions before the last one
until it hits a satisfiable sub-expression. Interleaving these ints speeds up
constraint solving in cases where a long list of addresses are returned and
only a few of them can satisfy constraints that are added later.