VariableManager: Only reuse phi vars for variables created in the same block.
This commit solves the issue that unrelated and unaccessed variables may
be added to an existing cluster of variables (governed by the same phi
variable). By limiting the reuse of phi variables, when an unrelated
variable U is added to an existing cluster of variables that already
have a phi variable A, we will create a new phi node B that governs
both A and U. This way it allows us to track if B is accessed or not
(instead of only tracking if A is accessed).