Include ins_addr in CodeLocation
Not having `ins_addr` in code location results in the same hash value for two code locations with the same statement idx but different ins_addr.
This situation arises when you are performing simplifications on a block that results in certain statements being removed, so the statement id now points to a different instruction.
The reason this is a problem is that when a new AST is created, claripy has a cache, which it checks to see if the AST has already been created. But since ins_addr is not included in the hash of the annotation, there is a possibility that a new AST with a different ins_addr but same stmt_idx already exists in the cache. So it returns this AST with the wrong ins_addr