Returning TOP in RDA for the whole register
This fix is for the following issue during RDA.
If a smaller part of a register is set first e.g. PUT(AX) = 0x5, then the whole register is retrieved later e.g. t0 = GET(RAX), t0 will now have TOP of size 64 instead of having TOP and the values 0x5 of respective sizes.
This happens because register_definitions.load() in _handle_Get() catches a `SimMemoryMissingError` which is raised in `default_filler_mixin.py` when it tries to fill the missing part because `fill_missing` is set to False always as a default. This exception is caught in _handle_Get() in RDA's vex engine, which creates a TOP of the size of the whole register instead of keeping the existing definitions for parts of the register.