Decompiler: Fix an infinite loop caused by type negotiation. (#3790)
CStructuredCodeGenerator._access() may end up in an infinite loop when
it tries to access offset 0 of any single-field structs. This is because
type negotiation converts the first field into the struct type itself.
As a temporary fix, we disallow type negotiation to return struct types
if the old type is not a struct type.