Allow Definiton tagging via FunctionEffect
After the RDA refactor a `FunctionHandler` should call
`FunctionCallData.depends` instead of modyifing the
`ReachingDefinitionsState` directly. But `.depends` didn't allow a `Tag` to be
passed, which made it impossible to use the tagging mechanism from
inside a `FunctionHandler` while adhering to the interface that a
`FunctionHandler` is expected to use.

This commit fixes this.