Factor out RDA State initialization logic
This commit factors out the logic that initializes the state at the
beginning of a ReachingDefinitionsAnalysis into a dedicated class.

It also splits out the steps of the default logic into substeps that can
easily be overridden in a subclass, with custom logic.

An instance of this class can then be passed when starting the RDA to
easily use any kind of more complicated logic to setup the state.