reassembler: Fix addition type error in address calculation
When an instruction contains an operand that is neither a memory
nor an immediate operand (e.g. a floating-point operand) then
`operand_offset` is set to `None`. This causes a type error when
calculating the address in `register_instruction_reference`:

https://github.com/angr/angr/blob/b0d099550163ca56424994c8ee80b8cacec6490c/angr/analyses/reassembler.py#L1980

I ran into this issue while packaging angr for Guix where the
`test_ln_gcc_O2` test case fails (due to this issue) with the
following error message:

	self = <Reassembler Analysis Result at 0x7fffe4c63190>, insn_addr = 4201104
	ref_addr = 6344696, sort = 'absolute', operand_offset = None

	    def register_instruction_reference(self, insn_addr, ref_addr, sort, operand_offset):
		if not self.log_relocations:
		    return

	>       addr = insn_addr + operand_offset
	E       TypeError: unsupported operand type(s) for +: 'int' and 'NoneType'

It seems to me that this regression was introduced in #1739, it is
presently unclear to me why this doesn't fail on the CI.