handle not finding initial blocks when simplifying stack canaries
StackCanarySimplifier._find_canary_init_stmt fails if it cannot find an
initial block as it calls `next(self.get_blocks(block_addr))` outside of
an iterator.

This PR handles the exception similarly to what happens if the next call
returns None.

The included integration test validates decompilation of a single
function from an existing binary failed prior to this fix.