fix: retry conditions are better respected for analyze_binaries
We've introduced a timeout to the analyze_binaries job to prevent the
accidental overuse of GitHub Actions minutes in the future.

The retries variable was missing and so the number of retries was
empty and never being evaluated.

Lastly, we update the gh_push_file.sh script to be smarter about
retriable conditions.