Fix Icicle breakpoints not reliably working in thumb mode (#5688)
Icicle does not set the thumb bit when considering to break for a
breakpoint. So for instance, if a breakpoint is set at 0x1001,
Icicle will never hit this, it would instead hit 0x1000. Since angr
prefers to pretend the thumb bit is set in the pc this makes sure it
is always unset before being communicated to Icicle.